+++ src/main/java/org/john_schreier/PRedictions/services/impl/UserStatsServiceImpl.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Service
 public class UserStatsServiceImpl implements UserStatsService {
-//    Implementation of the UserStats service which defines the logic for finding
+    //    Implementation of the UserStats service which defines the logic for finding
 //    a user's goals and history. Methods for updating and deleting goals/history have
 //    been defined for future implementation.
     @Autowired
@@ -27,6 +27,7 @@ public class UserStatsServiceImpl implements UserStatsService {
         return userStatsRepository.findAll();
     }
 
+    //  Method for users to save their coach choice, goals, and history.
     @Override
     public UserStats saveUserStats(UserStats userStats) throws PRException {
         if (userStats == null) {
@@ -43,12 +44,20 @@ public class UserStatsServiceImpl implements UserStatsService {
         return userStatsRepository.findByEmail(email);
     }
 
+    //  This method was added for future implementation, should a user want to delete their old goals and set new ones.
     @Override
-    public void deleteUserStatsByEmail(long id) {
+    public void deleteUserStatsByEmail(long id) throws PRException {
+        if (id == 0) {
+            throw new PRException("No UserStats found to delete.");
+        }
         userStatsRepository.deleteById(id);
     }
 
-    public void updateUserStats(UserStats userStats) {
+    //    This method was added for future implementation, should a user want to update their UserStats.
+    public void updateUserStats(UserStats userStats) throws PRException {
+        if (userStats == null) {
+            throw new PRException("No UserStats found for update.");
+        }
         userStatsRepository.save(userStats);
     }
